requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.1.0
nest-asyncio==1.6.0
outline-vpn-api==6.3.0
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
yarl==1.11.1
zipp==3.20.2
